// packages/sirius-web/frontend/sirius-web-application/src/views/edit-project/workbench-views/search/SearchView.types.ts
/*******************************************************************************
 * Copyright (c) 2025 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/

import { GQLSearchResult, SearchQuery } from './useSearch.types';

export interface SearchViewState {
  query: SearchQuery | null;
  result: GQLSearchResult | null;
  resultsReceivedTimestamp: number | null;
}